backend/requirements.txt
fastapi
uvicorn
python-multipart
python-jose[cryptography]
passlib[bcrypt]
httpx
chromadb
sentence-transformers
huggingface_hub
pydantic-settings
python-json-logger
websockets
slowapi
# New dependencies for file generation
reportlab
markdown
pypdf
python-docx
